Fix bug in fixing the openssl config file
Script no longer needs to be ran with sudo to fix config file, password
is prompted at the start of the script instead.

Only the editing of the file is running as root, rest of the script as
non-root.
